Move schema services to dedicated folder
