+++ frontend/src/services/schema/index.ts
@@ -1,0 +1,4 @@
+import FakeDataSchemaService from './FakeDataSchemaService';
+import SchemaService from './SchemaService';
+
+export { FakeDataSchemaService, SchemaService };
